+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased] - 2024-07-12
+## [2.0.0] - 2024-07-13
 
 ### Added
 - Users are now called to check their edu-ID credentials if a problem is detected during the login process.
@@ -45,7 +45,8 @@
 
 - Initial release.
 
-[Unreleased]: https://github.com/thechnet/adamatic/compare/v1.3.0...HEAD
+[Unreleased]: https://github.com/thechnet/adamatic/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/thechnet/adamatic/compare/v1.3.0...v2.0.0
 [1.3.0]: https://github.com/thechnet/adamatic/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/thechnet/adamatic/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/thechnet/adamatic/compare/v1.0.1...v1.1.0